Move InputDatum into own subnamespace
